@@ -6,15 +6,15 @@
 
     <groupId>ru.sliva</groupId>
     <artifactId>EasyChat</artifactId>
-    <version>1.0</version>
+    <version>2.0</version>
 
     <name>EasyChat</name>
 
     <description>Eazy chat for servers.</description>
 
     <properties>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
